+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 - Add support for default value on macros
+- Add support for test expression `instance of`, feature of [Craft CMS](https://craftcms.com/docs/5.x/reference/twig/tests.html#instance-of)
 
 ---
 ## 0.9.1 (2024-10-10)